Initial functionality
The daemon already does something at this point, by listening to the
pipe, and then trying to control the dunst daemon by sending the special
notifications through bdus to pause further notifications. The current
timeout is very short, so that obviously needs to change, and I want to
allow better control via the pipe, but this is a nice first step.