Initial working solution
Rewriting the logic basically from scratch. Now handles a single FIFO as
the input, while producing output that is intended for the i3 bar. This
means that the binary should be run by the bar, and then secondary
scripts can control the notifications by printing "pause" or "resume"
into the created FIFO (~/.dunst-pause).